Authorise recovery module adding/deleting with JWT authentication (#1527)
Migrates the suboptiomal authentication/authorisation for adding/removing recovery module addresses to the new JWT/SIWE-based approach:

- Remove `OnlySafeOwner`, `TimestampGuard`, `DisableRecoveryAlertsGuard` and `EnableRecoveryAlertsGuard` and associated tests
- Replace above guard usage `AuthGuard`
- Checksum incoming require Safe/module addresses
- Pass `AuthPayload` to domain and assert chain, signer and ownership before adding/deleting module
- Add/update associated test coverage

+++ src/routes/recovery/recovery.service.ts
@@ -1,21 +1,62 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, UnauthorizedException } from '@nestjs/common';
 import { IAlertsRepository } from '@/domain/alerts/alerts.repository.interface';
 import { AlertsRepository } from '@/domain/alerts/alerts.repository';
 import { AddRecoveryModuleDto } from '@/routes/recovery/entities/add-recovery-module.dto.entity';
 import { AlertsRegistration } from '@/domain/alerts/entities/alerts-registration.entity';
+import { AuthPayload } from '@/domain/auth/entities/auth-payload.entity';
+import { ISafeRepository } from '@/domain/safe/safe.repository.interface';
 
 @Injectable()
 export class RecoveryService {
   constructor(
     @Inject(IAlertsRepository)
     private readonly alertsRepository: AlertsRepository,
+    @Inject(ISafeRepository)
+    private readonly safeRepository: ISafeRepository,
   ) {}
 
   async addRecoveryModule(args: {
     chainId: string;
-    safeAddress: string;
+    safeAddress: `0x${string}`;
     addRecoveryModuleDto: AddRecoveryModuleDto;
+    authPayload: AuthPayload;
   }): Promise<void> {
+    if (
+      !args.authPayload.isForChain(args.chainId) ||
+      !args.authPayload.signer_address
+    ) {
+      throw new UnauthorizedException();
+    }
+
+    // Check after AuthPayload check to avoid unnecessary request
+    const isOwner = await this.safeRepository
+      .isOwner({
+        safeAddress: args.safeAddress,
+        chainId: args.chainId,
+        address: args.authPayload.signer_address,
+      })
+      // Swallow error to avoid leaking information
+      .catch(() => false);
+    if (!isOwner) {
+      throw new UnauthorizedException();
+    }
+
+    // After after owner check to avoid unnecessary request
+    const isEnabled = await this.safeRepository
+      .getSafesByModule({
+        chainId: args.chainId,
+        moduleAddress: args.addRecoveryModuleDto.moduleAddress,
+      })
+      .then(({ safes }) => {
+        return safes.some((safe) => safe === args.safeAddress);
+      })
+
+      // Swallow error to avoid leaking information
+      .catch(() => false);
+    if (!isEnabled) {
+      throw new UnauthorizedException();
+    }
+
     const contract: AlertsRegistration = {
       chainId: args.chainId,
       address: args.addRecoveryModuleDto.moduleAddress,
@@ -26,8 +67,30 @@ export class RecoveryService {
 
   async deleteRecoveryModule(args: {
     chainId: string;
-    moduleAddress: string;
+    moduleAddress: `0x${string}`;
+    safeAddress: `0x${string}`;
+    authPayload: AuthPayload;
   }): Promise<void> {
+    if (
+      !args.authPayload.isForChain(args.chainId) ||
+      !args.authPayload.signer_address
+    ) {
+      throw new UnauthorizedException();
+    }
+
+    // Check after AuthPayload check to avoid unnecessary request
+    const isOwner = await this.safeRepository
+      .isOwner({
+        safeAddress: args.safeAddress,
+        chainId: args.chainId,
+        address: args.authPayload.signer_address,
+      })
+      // Swallow error to avoid leaking information
+      .catch(() => false);
+    if (!isOwner) {
+      throw new UnauthorizedException();
+    }
+
     await this.alertsRepository.deleteContract({
       chainId: args.chainId,
       address: args.moduleAddress,
